+++ lib/bin/lwm2m_carrier/os/lwm2m_os.c
@@ -11,7 +11,7 @@
 #include <zephyr.h>
 #include <string.h>
 #include <modem/at_cmd.h>
-#include <at_notif.h>
+#include <modem/at_notif.h>
 #include <at_cmd_parser/at_cmd_parser.h>
 #include <at_cmd_parser/at_params.h>
 #include <bsd.h>
