+++ lib/modem_info/modem_info.c
@@ -6,7 +6,7 @@
 
 #include <at_cmd_parser/at_cmd_parser.h>
 #include <modem/at_cmd.h>
-#include <at_notif.h>
+#include <modem/at_notif.h>
 #include <ctype.h>
 #include <device.h>
 #include <errno.h>
