+++ samples/nrf9160/https_client/src/main.c
@@ -12,7 +12,7 @@
 #include <net/tls_credentials.h>
 #include <lte_lc.h>
 #include <modem/at_cmd.h>
-#include <at_notif.h>
+#include <modem/at_notif.h>
 #include <modem_key_mgmt.h>
 
 #define HTTPS_PORT 443
