+++ samples/nrf9160/serial_lte_modem/src/slm_at_host.c
@@ -12,7 +12,7 @@
 #include <string.h>
 #include <init.h>
 #include <modem/at_cmd.h>
-#include <at_notif.h>
+#include <modem/at_notif.h>
 
 LOG_MODULE_REGISTER(at_host, CONFIG_SLM_LOG_LEVEL);
